ReadFile and InsertBsonField refactory
